requirements.txt
selenium
pillow
requests
pathvalidate